requirements.txt
grpcio
grpcio-tools
mysql-connector-python
yfinance
cachetools
confluent-kafka
prometheus-client